requirements.txt
click==6.7
decorator==4.0.11
Flask==0.12.2
Flask-SQLAlchemy==2.2
ipython==6.1.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
MarkupSafe==1.0
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.5.2
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.11
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
